requirements.txt
Django==3.0.5
psycopg2-binary==2.8.5
djangorestframework==3.11.0
djangorestframework-jwt==1.11.0
django-rest-auth==0.9.5
requests==2.23.0
factory-boy==2.12.0
pytest==5.4.1
pytest-django==3.9.0